refactor: move common functionality into separate file
