+++ pkg/collectors/common.go
@@ -1,0 +1,17 @@
+package collectors
+
+import "github.com/prometheus/client_golang/prometheus"
+
+func collectGaugeValue(
+	ch chan<- prometheus.Metric,
+	desc *prometheus.Desc,
+	value float64,
+	labelValues []string,
+) {
+	ch <- prometheus.MustNewConstMetric(
+		desc,
+		prometheus.GaugeValue,
+		value,
+		labelValues...,
+	)
+}
